+++ libs/domain/product/category-related-list/category-related-list.schema.ts
@@ -1,0 +1,10 @@
+import { ContentComponentSchema } from '@spryker-oryx/experience';
+import { IconTypes } from '@spryker-oryx/ui/icon';
+import { ProductCategoryRelatedListComponent } from './category-related-list.component';
+
+export const productCategoryRelatedListSchema: ContentComponentSchema<ProductCategoryRelatedListComponent> =
+  {
+    name: 'Product Category Related List',
+    group: 'Product',
+    icon: IconTypes.ViewList,
+  };
