+++ libs/domain/product/category-related-list/index.ts
@@ -1,0 +1,2 @@
+export * from './category-related-list.component';
+export * from './category-related-list.schema';
