+++ libs/domain/product/list/src/list.model.ts
@@ -1,3 +1,3 @@
-import { ProductListQualifier } from '../../src/models';
+import { ProductListQualifier } from '@spryker-oryx/product';
 
 export type ProductListOptions = ProductListQualifier;
