+++ libs/domain/product/src/services/adapter/index.ts
@@ -1,7 +1,3 @@
-export * from './default-product-list.adapter';
-export * from './default-product-relations-list.adapter';
 export * from './default-product.adapter';
 export * from './normalizers';
-export * from './product-list.adapter';
-export * from './product-relations-list.adapter';
 export * from './product.adapter';
