+++ libs/domain/product/src/services/index.ts
@@ -1,14 +1,10 @@
 export * from './adapter';
-export * from './default-product-list-page.service';
-export * from './default-product-list.service';
-export * from './default-product-relations-list.service';
 export * from './default-product.service';
 export * from './images';
+export * from './list';
 export * from './product-context';
-export * from './product-list-page.service';
-export * from './product-list.service';
-export * from './product-relations-list.service';
 export * from './product.providers';
 export * from './product.service';
+export * from './related';
 export * from './resolvers';
 export * from './state';
