+++ libs/domain/product/src/services/list/adapter/default-product-list.adapter.spec.ts
@@ -2,10 +2,10 @@ import { HttpService, JsonAPITransformerService } from '@spryker-oryx/core';
 import { HttpTestService } from '@spryker-oryx/core/testing';
 import { createInjector, destroyInjector } from '@spryker-oryx/di';
 import { of } from 'rxjs';
-import { ProductQualifier } from '../../models/product-qualifier';
-import { DefaultProductRelationsListAdapter } from './default-product-relations-list.adapter';
-import { RelationsListNormalizer } from './normalizers/relations-list';
-import { ProductRelationsListAdapter } from './product-relations-list.adapter';
+import { ProductListQualifier } from '../../../models';
+import { ProductListNormalizer } from '../../adapter';
+import { DefaultProductListAdapter } from './default-product-list.adapter';
+import { ProductListAdapter } from './product-list.adapter';
 
 const mockApiUrl = 'mockApiUrl';
 const mockProducts = {
@@ -18,13 +18,12 @@ const mockProducts = {
     },
   ],
 };
-
 const mockTransformer = {
   do: vi.fn().mockReturnValue(() => of(null)),
 };
 
-describe('DefaultProductRelationsListAdapter', () => {
-  let adapter: ProductRelationsListAdapter;
+describe('DefaultProductCategoryAdapter', () => {
+  let service: ProductListAdapter;
   let http: HttpTestService;
 
   beforeEach(() => {
@@ -35,8 +34,8 @@ describe('DefaultProductRelationsListAdapter', () => {
           useClass: HttpTestService,
         },
         {
-          provide: ProductRelationsListAdapter,
-          useClass: DefaultProductRelationsListAdapter,
+          provide: ProductListAdapter,
+          useClass: DefaultProductListAdapter,
         },
         {
           provide: 'SCOS_BASE_URL',
@@ -49,23 +48,20 @@ describe('DefaultProductRelationsListAdapter', () => {
       ],
     });
 
-    adapter = testInjector.inject(
-      ProductRelationsListAdapter
-    ) as DefaultProductRelationsListAdapter;
+    service = testInjector.inject(ProductListAdapter);
     http = testInjector.inject(HttpService) as HttpTestService;
   });
 
   afterEach(() => {
-    vi.clearAllMocks();
     destroyInjector();
   });
 
   it('should be provided', () => {
-    expect(adapter).toBeInstanceOf(DefaultProductRelationsListAdapter);
+    expect(service).toBeInstanceOf(DefaultProductListAdapter);
   });
 
   describe('get method', () => {
-    const mockQualifier: ProductQualifier = { sku: 'test' };
+    const mockQualifier: ProductListQualifier = { q: 'test' };
 
     beforeEach(() => {
       http.flush(mockProducts);
@@ -75,17 +71,18 @@ describe('DefaultProductRelationsListAdapter', () => {
       vi.clearAllMocks();
     });
 
-    it('should have correct http path', () => {
-      adapter.get(mockQualifier);
+    it('should build correct base url', () => {
+      service.get(mockQualifier);
+
       expect(http.url).toContain(
-        `concrete-products/${mockQualifier.sku}/concrete-alternative-products`
+        `${mockApiUrl}/catalog-search?q=${mockQualifier.q}`
       );
     });
 
     it('should call transformer with proper normalizer', () => {
-      adapter.get(mockQualifier).subscribe();
+      service.get(mockQualifier).subscribe();
 
-      expect(mockTransformer.do).toHaveBeenCalledWith(RelationsListNormalizer);
+      expect(mockTransformer.do).toHaveBeenCalledWith(ProductListNormalizer);
     });
 
     it('should return transformed data', () => {
@@ -93,9 +90,25 @@ describe('DefaultProductRelationsListAdapter', () => {
       const callback = vi.fn();
       mockTransformer.do.mockReturnValue(() => of(mockTransformerData));
 
-      adapter.get(mockQualifier).subscribe(callback);
+      service.get(mockQualifier).subscribe(callback);
 
       expect(callback).toHaveBeenCalledWith(mockTransformerData);
+    });
+  });
+
+  describe('getKey method', () => {
+    it('should generate key from query string', () => {
+      const query = { q: 'test', brand: 'sony', color: 'red' };
+      expect(service.getKey(query)).toBe('q=test&brand=sony&color=red');
+    });
+
+    it('should correct transform alias to query string', () => {
+      const query = { q: 'test', maxPrice: 12, minPrice: 1 };
+      expect(service.getKey(query)).toBe('q=test&price[max]=12&price[min]=1');
+    });
+
+    it('should generate empty string when query param is not provided', () => {
+      expect(service.getKey({})).toBe('');
     });
   });
 });
