+++ libs/domain/product/src/services/list/adapter/index.ts
@@ -1,0 +1,2 @@
+export * from './default-product-list.adapter';
+export * from './product-list.adapter';
