+++ libs/domain/product/src/services/related/adapter/index.ts
@@ -1,0 +1,2 @@
+export * from './default-product-relations-list.adapter';
+export * from './product-relations-list.adapter';
