+++ libs/domain/product/src/services/related/index.ts
@@ -1,0 +1,3 @@
+export * from './adapter';
+export * from './default-product-relations-list.service';
+export * from './product-relations-list.service';
